Commit 2024-05-24

+++ ReachingFam.Core/Models/DataViewModels/ChartDataListModel.cs
@@ -8,6 +8,6 @@ namespace ReachingFam.Core.Models.DataViewModels
     public class ChartDataListModel
     {
         public List<string> Date { get; set; }
-        public List<double?> Visits { get; set; }
+        public List<decimal?> Visits { get; set; }
     }
 }
