+++ ReachingFam.Core/Models/ReachingFamViewModels/DashboardViewModel.cs
@@ -10,15 +10,15 @@ namespace ReachingFam.Core.Models.ReachingFamViewModels
     public class DashboardViewModel
     {
         public int FamilyCount { get; set; }
-        public double FamilyWeight { get; set;}
+        public decimal FamilyWeight { get; set;}
         public int SeniorCount { get; set; }
         public int AdultCount { get; set;}
         public int ChildrenCount { get; set;}
-        public double TotalHoursWorked { get; set;}
-        public double TotalWeightIn { get; set;}
-        public double TotalWeightOut { get; set;}
-        public double PartnerWeight { get; set;}
-        public double VolunteerWeight { get; set;}
+        public decimal TotalHoursWorked { get; set;}
+        public decimal TotalWeightIn { get; set;}
+        public decimal TotalWeightOut { get; set;}
+        public decimal PartnerWeight { get; set;}
+        public decimal VolunteerWeight { get; set;}
         public AnalyticPeriod Period { get; set;} = AnalyticPeriod.ThisMonth;
     }
