+++ ReachingFam.Core/Models/ReachingFamViewModels/PartnerGiveOutViewModel.cs
@@ -23,7 +23,7 @@ namespace ReachingFam.Core.Models.ReachingFamViewModels
         [DataType(DataType.Time)]
         public DateTime CollectionTime { get; set; }
         [Required]
-        public double Weight { get; set; }
+        public decimal Weight { get; set; }
         [Required]
         [Display(Name = "# Of Families")]
         public int NumberOfFamilies { get; set; }
@@ -33,17 +33,17 @@ namespace ReachingFam.Core.Models.ReachingFamViewModels
         [Display(Name = "Non Perishables")]
         public bool NonPerishables { get; set; } = false;
         [Display(Name = "NP Weight")]
-        public double? NonPerishablesWeight { get; set; }
+        public decimal? NonPerishablesWeight { get; set; }
         public bool Perishables { get; set; } = false;
         [Display(Name = "P Weight")]
-        public double? PerishablesWeight { get; set; }
+        public decimal? PerishablesWeight { get; set; }
         public bool Frozen { get; set; } = false;
         [Display(Name = "F Weight")]
-        public double? FrozenWeight { get; set; }
+        public decimal? FrozenWeight { get; set; }
         [Display(Name = "Non Food")]
         public bool NonFood { get; set; } = false;
         [Display(Name = "NF Weight")]
-        public double? NonFoodWeight { get; set; }
+        public decimal? NonFoodWeight { get; set; }
         [Display(Name = "Collected?")]
         public bool Collected { get; set; } = false;
         [Display(Name = "Date Collected")]
