+++ ReachingFam.Core/Models/ReachingFamViewModels/VolunteerGiveOutViewModel.cs
@@ -18,20 +18,20 @@ namespace ReachingFam.Core.Models.ReachingFamViewModels
         [Display(Name = "Collection Date")]
         public DateTime CollectionDate { get; set; }
         [Required]
-        public double Weight { get; set; }
+        public decimal Weight { get; set; }
         [Display(Name = "Non Perishables")]
         public bool NonPerishables { get; set; } = false;
         [Display(Name = "NP Weight")]
-        public double? NonPerishablesWeight { get; set; }
+        public decimal? NonPerishablesWeight { get; set; }
         public bool Perishables { get; set; } = false;
         [Display(Name = "P Weight")]
-        public double? PerishablesWeight { get; set; }
+        public decimal? PerishablesWeight { get; set; }
         public bool Frozen { get; set; } = false;
         [Display(Name = "F Weight")]
-        public double? FrozenWeight { get; set; }
+        public decimal? FrozenWeight { get; set; }
         [Display(Name = "Non Food")]
         public bool NonFood { get; set; } = false;
         [Display(Name = "NF Weight")]
-        public double? NonFoodWeight { get; set; }
+        public decimal? NonFoodWeight { get; set; }
     }
 }
